fix pageworklaod (#5165)
close pingcap/tiflash#5164
