+++ dbms/src/Storages/Page/workload/Normal.h
@@ -77,6 +77,4 @@ public:
         stop_watch.stop();
     }
 };
-
-REGISTER_WORKLOAD(NormalWorkload)
 } // namespace DB::PS::tests
